Add `captureR` message type
